requirements.txt
numpy==1.25.2
pip==23.2.1
pybullet==3.2.5
setuptools==65.5.0
PyYAML==6.0.1
ipython==8.16.1
matplotlib==3.8.0
tensorboard==2.14.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
packaging==23.1
pandas==2.1.2